All working with flattened data, but React warnings appear in console.
The handling of re-shaping the data so that it works with the existing component structure has resulted in warnings in the console from React instructing not to set props in the components but to do so when the data is received. This is probably a cleaner option given that the re-shaping would take place all in one file.