Avoid React warnings by not altering props and moving functionality
Rather than defining this.props.whatever inside of the component, I called the funcitons which would have assigned the reshaped object/array to this.props.whatever from within the render method and used a variable with the value of whatever those reshapers returned instead of this.props.monthData or this.props.dayData. I'm not sure if this is cheating or breaking some principal, but it seems to be working for the time being.