Rework components so state is altered instead of setting prop values
Reworked Top-Level components so that when the component is first loading or when new props are about to be loaded the incoming values are restructured to fit the component hierarchy and then set as the current state before the components are rendered.

I tried to commit these changes in a separate branch, but dealt with a bunch of nonsense stemming from my uploading a file from work into master, the craziness of bundle.js, etc. In any event, all the changes I'd tried to include in the other branch are here.