Update ScheduleStore.js to better handle Weekend shift assignment
Updated ScheduleStore.js to more efficiently sort through employees to find a viable candidate for assignment to a shift. Updated code so "Unassigned" is not treated as an employee for purposes of shift assignment. Fixed code that determines what the lowest number of shifts any employee has at a given momentis returning the fewest shifts and not most shifts.